// PremiereInterna-master/src/Models/IFuncionario.ts
export interface IFuncionario {
  id?: number;
  nome?: string;
  cpf?: string;
  rg?: string;
  data_nascimento?: any;
  endereco?: string;
  cep?: string;
  bairro?: string;
  cidade?: string;
  estado?: string;
  telefone?: string;
  email?: string;
  funcao?: string;
  login?: string;
  senha?: null;
}